pyproject.toml
[project]
name = "test-competencias"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.5",
    "djangorestframework>=3.15.2",
    "pytest>=8.3.4",
    "pytest-django>=4.9.0",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.10",
    "requests>=2.32.3",
]
